move getThemeLogo to themeUtils
