+++ src/utils/themeUtils.ts
@@ -15,7 +15,41 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { Theme, THEMES } from "@/options/constants";
+import { DEFAULT_THEME, Theme, THEMES } from "@/options/constants";
+import logo from "@img/logo.svg";
+import logoSmall from "@img/logo-small-rounded.svg";
+import aaLogo from "@img/aa-logo.svg";
+import aaLogoSmall from "@img/aa-logo-small.svg";
 
 export const isValidTheme = (theme: string): theme is Theme =>
   THEMES.includes(theme as Theme);
+
+export type ThemeLogo = {
+  regular: string;
+  small: string;
+};
+
+type ThemeLogoMap = {
+  [key in Theme]: ThemeLogo;
+};
+
+const THEME_LOGOS: ThemeLogoMap = {
+  default: {
+    regular: logo,
+    small: logoSmall,
+  },
+  "automation-anywhere": {
+    regular: aaLogo,
+    small: aaLogoSmall,
+  },
+};
+
+export const getThemeLogo = (theme: string): ThemeLogo => {
+  if (isValidTheme(theme)) {
+    // eslint-disable-next-line security/detect-object-injection -- theme is type Theme, a union type of string literal
+    return THEME_LOGOS[theme];
+  }
+
+  // eslint-disable-next-line security/detect-object-injection -- theme not user defined
+  return THEME_LOGOS[DEFAULT_THEME];
+};
